--- a/Backend/Solver/CrankNicolson/Iterate.c
+++ b/Backend/Solver/CrankNicolson/Iterate.c
@@ -71,7 +71,7 @@ Complex **Iterate1d(Function potential, Function psi0, int n)
         res[i] = matrixToArray(psi);
         // A = (I - dt/2 * (D2 + V))
         Matrix f = factorIpTri(sumIpTri(potentialMatrix, d2), (Complex) {.real = dt/2, .imaginary = 0});
-        Matrix a = subtract(ident, f);
+        Matrix a = subtractTri(ident, f);
         // b = (I + dt/2 * (D2 + V)) * Psi
         Matrix b = multiplyTri(sumIpTri(f, ident), psi);