--- a/Backend/Solver/Matrix/Utility.c
+++ b/Backend/Solver/Matrix/Utility.c
@@ -116,6 +116,25 @@ Matrix sum(Matrix a, Matrix b)
     return result;
 }
 
+Matrix sumIp(Matrix a, Matrix b)
+{
+    if(a.rowCount != b.rowCount || a.columnCount != b.columnCount){
+        fprintf(stderr, "Error: a and b have different dimensions");
+        exit(2);
+    }
+    int r = a.rowCount;
+    int c = a.columnCount;
+    Matrix result = a;
+    int i,j;
+    for(i = 0; i < r ; i++){
+        for(j = 0; j < c; j++){
+            result.matrix[i][j] = sumComplex(a.matrix[i][j], b.matrix[i][j]);
+        }
+    }
+
+    return result;
+}
+
 Matrix subtract(Matrix a, Matrix b)
 {
     if(a.rowCount != b.rowCount || a.columnCount != b.columnCount){