--- a/Backend/Solver/Matrix/Utility.h
+++ b/Backend/Solver/Matrix/Utility.h
@@ -19,6 +19,10 @@ Complex getElement(const Matrix m, const int r, const int c);
 Matrix multiply(Matrix a, Matrix b);
 Matrix factor(Matrix a, Complex f);
 Matrix sum(Matrix a, Matrix b);
+/**
+ * Result will be stored in a
+ */
+Matrix sumIp(Matrix a, Matrix b);
 Matrix subtract(Matrix a, Matrix b);
 Matrix arrayToMatrix(Complex *array, int n);
 Complex *matrixToArray(Matrix m);