--- a/Backend/Solver/Matrix/Utility.h
+++ b/Backend/Solver/Matrix/Utility.h
@@ -21,16 +21,17 @@ Matrix multiply(Matrix a, Matrix b);
  * Assumes a is tridiagonal
  */
 Matrix multiplyTri(Matrix a, Matrix b);
+// TODO: Not only support nxn
 Matrix factor(Matrix a, Complex f);
 /**
- * Result will be stored in a
+ * Result will be stored in a and assumes a is tridigaonal
  */
-Matrix factorIp(Matrix a, Complex f);
+Matrix factorIpTri(Matrix a, Complex f);
 Matrix sum(Matrix a, Matrix b);
 /**
- * Result will be stored in a
+ * Result will be stored in a. Assumes a and b are tridiagonal
  */
-Matrix sumIp(Matrix a, Matrix b);
+Matrix sumIpTri(Matrix a, Matrix b);
 Matrix subtract(Matrix a, Matrix b);
 Matrix arrayToMatrix(Complex *array, int n);
 Complex *matrixToArray(Matrix m);