--- a/run.sh
+++ b/run.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+Backend/Program.x $1 $2
+cd Frontend/src
+java de.numericat.Main "../../output"