Improved performance by 917% - before:

    19.34s user 0.55s system 99% cpu 19.983 total

after:

    1.77s user 0.39s system 98% cpu 2.180 total

\- by implementing in place and tridiagonal matrix operations and only evaluating the time dependent potential if it actually is time dependent.